Update calculating brightness & saturation indices
As per the research paper followed for building this algorithm (see: http://hplabs.hp.com/techreports/2005/HPL-2005-14.pdf), saturation index and brightness index are defined as: 
Saturation Index: To implement this, we compute the color saturation and subtracted the average color saturation of the blurry areas from that of the sharp areas. 
Brightness Index: To implement this, we compute the average brightness difference between the sharp and blurry areas. 
Proposed relevant changes required to calculate brightness and saturation index in this regard.